+++ Changelog.md
@@ -16,6 +16,7 @@ Changes
 - Minimum max-age for HSTS is now 14 months. [(#421)]
 - Accept all 3xx+3xx and 3xx+2xx DANE rollover schemes. [(#341)]
 - Certificate Usage Field on TLSA records for email test. [(#329)]
+- Validate CSP directives. [(#325)]
 
 Bug Fixes
 - Fix indefinite locks in cache (not a current problem).
@@ -41,8 +42,9 @@ Settings
 - New SMTP_EHLO_DOMAIN setting in settings.py. [(#483)]
 
 [(#249)]: https://github.com/NLnetLabs/Internet.nl/issues/249
-[(#341)]: https://github.com/NLnetLabs/Internet.nl/issues/341
 [(#329)]: https://github.com/NLnetLabs/Internet.nl/issues/329
+[(#325)]: https://github.com/NLnetLabs/Internet.nl/issues/325
+[(#341)]: https://github.com/NLnetLabs/Internet.nl/issues/341
 [(#421)]: https://github.com/NLnetLabs/Internet.nl/issues/421
 [(#443)]: https://github.com/NLnetLabs/Internet.nl/issues/443
 [(#461)]: https://github.com/NLnetLabs/Internet.nl/issues/461